requirements.txt
ascii_magic
datetime
python-Levenshtein
pathlib
pexpect
tf-keras
beautifulsoup4
chronos
climage
clipboard
customtkinter
deepface
dnspython
elasticsearch
Flask
gitignore_parser
google_api_python_client
google_auth_oauthlib
groq
gTTS
halo
html2text
huggingface_hub
InquirerPy
matplotlib
mss
nltk
numpy
ollama
openai
opencv_python_headless
pandas
piexif
Pillow
plotly
prompt_toolkit
PyAudio
pydub
pyenchant
pyfiglet
pygame
pygame_ce
Pygments
pynput
pytesseract
python_dateutil
python_magic
pyttsx3
qrcode
requests
rich
scikit_learn
scipy
selenium
simpleaudio
sounddevice
spacy
SpeechRecognition
sumy
tensorflow
pytesseract
pdfplumber
python-docx
textstat
thefuzz
tiktoken
torch
tqdm
transformers
typing_extensions
webcolors
webdriver_manager
youtube_dl
youtube_transcript_api
yt_dlp
frontend
docker
pgeocode
netifaces
dnspython
markdownify
tabulate
psutil
playwright
